project setup and refactoring
