--- src/main/java/khirtech/coronadz/demo/entities/User.java
@@ -1,4 +1,0 @@
-package khirtech.coronadz.demo.entities;
-
-public class User {
-}
