+++ src/main/java/khirtech/coronadz/demo/infection/Infection.java
@@ -1,4 +1,4 @@
-package khirtech.coronadz.demo.entities;
+package khirtech.coronadz.demo.infection;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
