+++ src/main/java/khirtech/coronadz/demo/infection/InfectionController.java
@@ -1,6 +1,5 @@
 package khirtech.coronadz.demo.infection;
 
-import khirtech.coronadz.demo.entities.Infection;
 import org.springframework.data.mongodb.core.geo.GeoJsonPoint;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
