+++ src/main/java/khirtech/coronadz/demo/infection/InfectionRepository.java
@@ -1,6 +1,5 @@
 package khirtech.coronadz.demo.infection;
 
-import khirtech.coronadz.demo.entities.Infection;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
